+++ pkg/providers/papi/ruleformats.gen_test.go
@@ -402,6 +402,26 @@ func TestProcessPolicyTemplatesGenerated(t *testing.T) {
 			filesToCheck: []string{"property.tf", "rules.tf", "variables.tf", "import.sh"},
 			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2026-01-09")},
 		},
+		"property with rules as datasource - hcl rules version v2026-02-16": {
+			givenData: TFData{
+				Property: TFPropertyData{
+					GroupName:            "test_group",
+					GroupID:              "grp_12345",
+					ContractID:           "test_contract",
+					PropertyResourceName: "test-edgesuite-net",
+					PropertyName:         "test.edgesuite.net",
+					PropertyID:           "prp_12345",
+					ProductID:            "prd_HTTP_Content_Del",
+					ProductName:          "HTTP_Content_Del",
+					RuleFormat:           "v2026-02-16",
+					IsSecure:             "false",
+					ReadVersion:          "LATEST",
+				},
+			},
+			dir:          "ruleformats/v2026_02_16/rules_datasource",
+			filesToCheck: []string{"property.tf", "rules.tf", "variables.tf", "import.sh"},
+			filterFuncs:  []func([]string) ([]string, error){useThisOnlyRuleFormat("v2026-02-16")},
+		},
 	}
 
 	for name, test := range tests {